refactor: set root directory + rename image-resource directory into assets
