+++ src/components/ProductReviews.jsx
@@ -1,7 +1,7 @@
 import { useState } from 'react';
 import Button from './Button';
-import emptyInquiryImg from '../image-resource/Img-inquiry-empty.svg';
-import kebabIcon from '../image-resource/ic-kebab.svg';
+import emptyInquiryImg from '../assets/img-inquiry-empty.svg';
+import kebabIcon from '../assets/ic-kebab.svg';
 
 const PLACEHOLDER =
   '개인정보를 공유 및 요청하거나, 명예 훼손, 무단 광고, 불법 정보 유포시 모니터링 후 삭제될 수 있으며, 이에 대한 민형사상 책임은 게시자에게 있습니다.';
